ArgusAPI/requirements.txt
django
django-filter
django-phonenumber-field[phonenumbers]
djangorestframework 
django-debug-toolbar
django-cors-headers
redis
celery
dlib
face-recognition
opencv-python
ffmpeg
numpy
pytz
whitenoise
mysqlclient
mysql-connector-python
